--- a/benches/bench.rs
+++ b/benches/bench.rs
@@ -46,7 +46,8 @@ fn write_b_bench(b: &mut Bencher, tx_size: TxSize, qindex: usize) {
   let mut w = ec::WriterEncoder::new();
   let mut fc = CDFContext::new(fi.base_q_idx);
   let mut fb = FrameBlocks::new(fi.sb_width * 16, fi.sb_height * 16);
-  let bc = BlockContext::new(&mut fb);
+  let mut tb = fb.as_tile_blocks_mut();
+  let bc = BlockContext::new(&mut tb);
   let mut fs = FrameState::new(&fi);
   let mut ts = fs.as_tile_state_mut();
   // For now, restoration unit size is locked to superblock size.
@@ -116,11 +117,10 @@ fn cdef_frame_bench(b: &mut Bencher, width: usize, height: usize) {
   };
   let sequence = Sequence::new(&Default::default());
   let fi = FrameInvariants::<u16>::new(config, sequence);
-  let mut fb = FrameBlocks::new(fi.sb_width * 16, fi.sb_height * 16);
-  let bc = BlockContext::new(&mut fb);
+  let fb = FrameBlocks::new(fi.sb_width * 16, fi.sb_height * 16);
   let mut fs = FrameState::new(&fi);
 
-  b.iter(|| cdef_filter_frame(&fi, &mut fs.rec, &bc.blocks));
+  b.iter(|| cdef_filter_frame(&fi, &mut fs.rec, &fb));
 }
 
 fn cfl_rdo(c: &mut Criterion) {