--- a/benches/predict.rs
+++ b/benches/predict.rs
@@ -88,7 +88,7 @@ pub fn intra_dc_4x4(b: &mut Bencher) {
   let (mut block, above, left) = generate_block(&mut rng);
 
   b.iter(|| {
-    Block4x4::pred_dc(&mut block.as_mut_slice(), &above[..4], &left[..4]);
+    Block4x4::pred_dc(&mut block.as_region_mut(), &above[..4], &left[..4]);
   })
 }
 
@@ -98,7 +98,7 @@ pub fn intra_dc_left_4x4(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_dc_left(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[..4]
     );
@@ -111,7 +111,7 @@ pub fn intra_dc_top_4x4(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_dc_top(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[..4]
     );
@@ -123,7 +123,7 @@ pub fn intra_h_4x4(b: &mut Bencher) {
   let (mut block, _above, left) = generate_block(&mut rng);
 
   b.iter(|| {
-    Block4x4::pred_h(&mut block.as_mut_slice(), &left[..4]);
+    Block4x4::pred_h(&mut block.as_region_mut(), &left[..4]);
   })
 }
 
@@ -132,7 +132,7 @@ pub fn intra_v_4x4(b: &mut Bencher) {
   let (mut block, above, _left) = generate_block(&mut rng);
 
   b.iter(|| {
-    Block4x4::pred_v(&mut block.as_mut_slice(), &above[..4]);
+    Block4x4::pred_v(&mut block.as_region_mut(), &above[..4]);
   })
 }
 
@@ -143,7 +143,7 @@ pub fn intra_paeth_4x4(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_paeth(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[..4],
       above_left
@@ -157,7 +157,7 @@ pub fn intra_smooth_4x4(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_smooth(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[..4]
     );
@@ -170,7 +170,7 @@ pub fn intra_smooth_h_4x4(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_smooth_h(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[..4]
     );
@@ -183,7 +183,7 @@ pub fn intra_smooth_v_4x4(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_smooth_v(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[..4]
     );
@@ -198,7 +198,7 @@ pub fn intra_cfl_4x4(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_cfl(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &ac,
       alpha,
       8,
@@ -215,7 +215,7 @@ pub fn intra_dc_4x4_u8(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_dc(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[32 - 4..]
     );
@@ -228,7 +228,7 @@ pub fn intra_dc_128_4x4_u8(b: &mut Bencher) {
   let (mut block, _above, _left) = generate_block_u8(&mut rng, &mut edge_buf);
 
   b.iter(|| {
-    Block4x4::pred_dc_128(&mut block.as_mut_slice(), 8);
+    Block4x4::pred_dc_128(&mut block.as_region_mut(), 8);
   })
 }
 
@@ -239,7 +239,7 @@ pub fn intra_dc_left_4x4_u8(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_dc_left(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[32 - 4..]
     );
@@ -253,7 +253,7 @@ pub fn intra_dc_top_4x4_u8(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_dc_top(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[32 - 4..]
     );
@@ -266,7 +266,7 @@ pub fn intra_h_4x4_u8(b: &mut Bencher) {
   let (mut block, _above, left) = generate_block_u8(&mut rng, &mut edge_buf);
 
   b.iter(|| {
-    Block4x4::pred_h(&mut block.as_mut_slice(), &left[32 - 4..]);
+    Block4x4::pred_h(&mut block.as_region_mut(), &left[32 - 4..]);
   })
 }
 
@@ -276,7 +276,7 @@ pub fn intra_v_4x4_u8(b: &mut Bencher) {
   let (mut block, above, _left) = generate_block_u8(&mut rng, &mut edge_buf);
 
   b.iter(|| {
-    Block4x4::pred_v(&mut block.as_mut_slice(), &above[..4]);
+    Block4x4::pred_v(&mut block.as_region_mut(), &above[..4]);
   })
 }
 
@@ -288,7 +288,7 @@ pub fn intra_paeth_4x4_u8(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_paeth(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[32 - 4..],
       above_left
@@ -303,7 +303,7 @@ pub fn intra_smooth_4x4_u8(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_smooth(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[32 - 4..]
     );
@@ -317,7 +317,7 @@ pub fn intra_smooth_h_4x4_u8(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_smooth_h(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[32 - 4..]
     );
@@ -331,7 +331,7 @@ pub fn intra_smooth_v_4x4_u8(b: &mut Bencher) {
 
   b.iter(|| {
     Block4x4::pred_smooth_v(
-      &mut block.as_mut_slice(),
+      &mut block.as_region_mut(),
       &above[..4],
       &left[32 - 4..]
     );