--- a/src/cdef.rs
+++ b/src/cdef.rs
@@ -206,7 +206,7 @@ fn adjust_strength(strength: i32, var: i32) -> i32 {
 
 pub fn cdef_analyze_superblock<T: Pixel>(
   in_frame: &Frame<T>,
-  blocks: &FrameBlocks,
+  blocks: &TileBlocks<'_>,
   sbo: SuperBlockOffset,
   sbo_global: SuperBlockOffset,
   bit_depth: usize,
@@ -221,7 +221,7 @@ pub fn cdef_analyze_superblock<T: Pixel>(
       // single-SB copy 'frame' that represents some superblock
       // in the main frame.
       let global_block_offset = sbo_global.block_offset(bx<<1, by<<1);
-      if global_block_offset.x < blocks.cols && global_block_offset.y < blocks.rows {
+      if global_block_offset.x < blocks.cols() && global_block_offset.y < blocks.rows() {
         let skip = blocks[global_block_offset].skip
           & blocks[sbo_global.block_offset(2*bx+1, 2*by)].skip
           & blocks[sbo_global.block_offset(2*bx, 2*by+1)].skip
@@ -357,7 +357,7 @@ pub fn cdef_filter_superblock<T: Pixel>(
   fi: &FrameInvariants<T>,
   in_frame: &Frame<u16>,
   out_frame: &mut Frame<T>,
-  blocks: &FrameBlocks,
+  blocks: &TileBlocks<'_>,
   sbo: SuperBlockOffset,
   sbo_global: SuperBlockOffset,
   cdef_index: u8,
@@ -382,7 +382,7 @@ pub fn cdef_filter_superblock<T: Pixel>(
   for by in 0..8 {
     for bx in 0..8 {
       let global_block_offset = sbo_global.block_offset(bx<<1, by<<1);
-      if global_block_offset.x < blocks.cols && global_block_offset.y < blocks.rows {
+      if global_block_offset.x < blocks.cols() && global_block_offset.y < blocks.rows() {
         let skip = blocks[global_block_offset].skip
           & blocks[sbo_global.block_offset(2*bx+1, 2*by)].skip
           & blocks[sbo_global.block_offset(2*bx, 2*by+1)].skip
@@ -501,13 +501,15 @@ pub fn cdef_filter_frame<T: Pixel>(fi: &FrameInvariants<T>, rec: &mut Frame<T>, 
     }
   }
 
+  let tb = blocks.as_tile_blocks();
+
   // Perform actual CDEF, using the padded copy as source, and the input rec vector as destination.
   for fby in 0..fb_height {
     for fbx in 0..fb_width {
       let sbo = SuperBlockOffset { x: fbx, y: fby };
       let cdef_index = blocks[sbo.block_offset(0, 0)].cdef_index;
-      let cdef_dirs = cdef_analyze_superblock(&cdef_frame, blocks, sbo, sbo, fi.sequence.bit_depth);
-      cdef_filter_superblock(fi, &cdef_frame, rec, blocks, sbo, sbo, cdef_index, &cdef_dirs);
+      let cdef_dirs = cdef_analyze_superblock(&cdef_frame, &tb, sbo, sbo, fi.sequence.bit_depth);
+      cdef_filter_superblock(fi, &cdef_frame, rec, &tb, sbo, sbo, cdef_index, &cdef_dirs);
     }
   }
 }