--- a/src/encoder.rs
+++ b/src/encoder.rs
@@ -383,6 +383,16 @@ impl Sequence {
       second_forward_idx >= 0
     }
   }
+
+  #[inline(always)]
+  pub fn sb_size_log2(&self) -> usize {
+    if self.use_128x128_superblock { 7 } else { 6 }
+  }
+
+  #[inline(always)]
+  pub fn sb_size(&self) -> usize {
+    1 << self.sb_size_log2()
+  }
 }
 
 #[derive(Debug)]
@@ -835,6 +845,16 @@ impl<T: Pixel> FrameInvariants<T> {
     self.lambda =
       qps.lambda * ((1 << (2 * (self.sequence.bit_depth - 8))) as f64);
     self.me_lambda = self.lambda.sqrt();
+  }
+
+  #[inline(always)]
+  pub fn sb_size_log2(&self) -> usize {
+    self.sequence.sb_size_log2()
+  }
+
+  #[inline(always)]
+  pub fn sb_size(&self) -> usize {
+    self.sequence.sb_size()
   }
 }