_(description updated on 16 april 2019)_

This PR implements tile encoding (#631).

 1. The (many) first commits introduce tiling structures, which allow to expose simultaneous tiled regions of the whole frame data.
 2. Following commits use the tiling structures where necessary in the whole codebase.
 3. Then, command line arguments are added, the encoder encodes tiles separately, and separate tiles are written to the bitstream.
 4. Finally, parallelization is enabled (spoiler: 1 line of code).


## Context

Encoding a frame first involves frame-wise accesses (initialization, etc.), then tile-wise accesses (to encode tiles in parallel), then frame-wise accesses using the results of tile-encoding (deblocking, cdef, …):

```
                                \
      +----------------+         |
      |                |         |
      |                |         |  Frame-wise accesses
      |                |          >
      |                |         |   - FrameState<T>
      |                |         |   - Frame<T>
      +----------------+         |   - Plane<T>
                                /    - ...

              ||   tiling views
              \/
                                \
  +---+  +---+  +---+  +---+     |
  |   |  |   |  |   |  |   |     |  Tile encoding (possibly in parallel)
  +---+  +---+  +---+  +---+     |
                                 |
  +---+  +---+  +---+  +---+     |  Tile-wise accesses
  |   |  |   |  |   |  |   |      >
  +---+  +---+  +---+  +---+     |   - TileStateMut<'_, T>
                                 |   - TileMut<'_, T>
  +---+  +---+  +---+  +---+     |   - PlaneRegionMut<'_, T>
  |   |  |   |  |   |  |   |     |
  +---+  +---+  +---+  +---+     |
                                /

              ||   vanishing of tiling views
              \/
                                \
      +----------------+         |
      |                |         |
      |                |         |  Frame-wise accesses
      |                |          >
      |                |         |  (deblocking, cdef, ...)
      |                |         |
      +----------------+         |
                                /
```



## Tiling

As you know, in Rust, it is not sufficient not to read/write the same memory from several threads, it must be impossible to write (safe) code that could do it. More precisely, a mutable reference may not alias any other reference to the same memory.

That's the reason why, as a preliminary step, I replaced accesses using the whole plane as a raw slice in addition to the stride information by `PlaneSlice` (#1035) and `PlaneMutSlice` (#1043).

But `Plane(Mut)Slice` still borrows the whole plane slice, so it does not, in itself, solves the problem.


There are several structures to be tiled, which form a tree:

```
 +- FrameState → TileState
 |  +- Frame → Tile
 |  |  +- Plane → PlaneRegion 
 |  +  RestorationState → TileRestorationState
 |  |  +- RestorationPlane → TileRestorationPlane
 |  |     +- FrameRestorationUnits → TileRestorationUnits
 |  +  FrameMotionVectors → TileMotionVectors
 +- FrameBlocks → TileBlocks
```

Most of them exist both in const and mutable version (e.g. `PlaneRegion` and `PlaneRegionMut`).


### Tiling structures

#### `PlaneRegion`

This is a view of bounded region of a `Plane`. It is similar to `PlaneSlice`, except that it does not borrow the whole underlying raw slice. That way, it is possible to get several non-overlapping regions simultaneously.

_In the end, we should probably merge it with `PlaneSlice`, but it requires more work because some frame-wise code still uses `PlaneSlice` in the code base._

It is possible to retrieve a subregion of a region (which may not exceed its parent). In theory, a subregion is defined by a rectangle (for example: x, y, width, height), but in practice, we need more flexibility. For example, we often need to retrieve a region from an offset, using the same bottom-right corner as its parent without providing width and height.

For that purpose, I propose a specific `Area` structure (actually, a Rust `enum`) to describe subregion bounds. Here are some usage examples:

```rust
let region = plane.region(Area::Rect { x: 32, y: 32, width: 512, height: 512 });

// the area is relative to the parent region
let subregion = region.subregion(Area::StartingAt { x: 128, y: 128 });
// it is equivalent to
let subregion = region.subregion(Area::Rect { x: 128, y: 128, width: 384, height: 384 });
// or
let subregion = plane.region(Area:: Rect { x: 160, y: 160, width: 384, height: 384 });
```

Retrieving a subregion from a `BlockOffset` is so common accross the code base that I decided to expose it directly:

```
let bo = BlockOffset { x: 2, y: 3 };
let subregion = region.subregion(Area::BlockStartingAt { bo });
```

Like `Plane(Mut)Slice`, it provides _operator[]_ and iterators over its rows:

```rust
let row5 = &region[5];
let value = region[3][4];
for row in region.rows_iter() {
    let _first_four_values = &row[..4];
}
```

The mutable versions of the structure (`PlaneRegionMut`) and methods are also provided.


#### `Tile`

A `Tile` is a view of 3 colocated plane regions (`Tile` is to a `PlaneRegion` as a `Frame` is to a `Plane`).

The mutable version (`TileMut`) is also provided.


#### `TileState`

The way the `FrameState` fields are mapped in `TileState` depends on how they are accessed tile-wise and frame-wise.

Some fields (like `qc`) are only used during tile-encoding, so they are only stored in `TileState`.

Some other fields (like `input` or `segmentation`) are not written tile-wise, so they just reference the matching field in `FrameState`.

Some others (like `rec`) are written tile-wise, but must be accessible frame-wise once the tile views vanish (e.g. for deblocking).

It contains 2 tiled views: `TileRestorationState` and a vector of `TileMotionVectorsMut` (a tiled view of `FrameMotionVectors`).

This structure is only provided as mutable (`TileStateMut`). A const version is not necessary, and would require to instantiate a const version of all its embedded tiled views.


#### `TileBlocks`

`TileBlocks` is a tiled view of `FrameBlocks`. It exposes the blocks associated to the tile.

The mutable version (`TileBlocksMut`) is also provided.


### Splitting into tiles

A `TilingInfo` structure computes all the details about tiling from the frame width and height and the (log2 of the) number of tile columns and rows. The details are accessible for initializing data or writing into the bitstream.

It provides an iterator over tiles (yielding one `TileStateMut` and one `TileBlocksMut` for each tile).

## Frame offsets vs tile offsets

In `encode_tile()`, super-block, block and plane offsets are expressed relative to the tile. The tiling views expose its data relative to the tile:
 - `plane_region[y][x]` is pixel (x, y) relative to the plane region,
 - `tile_blocks[boy][box]` contains the `Block` at (box, boy) relative to the tile,
 - …

`TileStateMut` exposes some references to frame-level data stored in `FrameState`:
 - `input` is a reference to the whole frame,
 - `input_hres` and `input_qres` are references to the whole planes.

When accessing these frame-level data, tile offsets are converted to frame offsets, for example by:

```rust
let frame_bo = ts.to_frame_block_offset(bo);
```


## Current state

It works.

Need more tests and reviews.