@rom1v 2019-03-19
I added `#[derive(Copy)]` to `BlockOffset`, and used `BlockOffset` by value in the `Area` enum. Using a reference would add a lifetime parameter (`Area<'_>`), which would be overkill here.

But since I added `Copy`, _clippy_ warns about `BlockOffset` everywhere:

```
error: this argument is passed by reference, but would be more efficient if passed by value
    --> src/partition.rs:1234:19
     |
1234 | pub fn has_tr(bo: &BlockOffset, bsize: BlockSize) -> bool {
     |                   ^^^^^^^^^^^^ help: consider passing by value instead: `BlockOffset`
     |
     = note: `-D clippy::trivially-copy-pass-by-ref` implied by `-D warnings`
     = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#trivially_copy_pass_by_ref
```

I think we should replace all `&BlockOffset` to `BlockOffset` in function parameters. Do you agree I do this on `master`?