@rom1v 2019-04-05
Here is my first working tiled (2×2) video encoded with rav1e: [bbb_tiled3.mp4](http://h.rom1v.com/bbb_tiled3.mp4) | [bbb_tiled3.ivf](http://h.rom1v.com/bbb_tiled3.ivf)

From this version: https://github.com/rom1v/rav1e/commits/tiling.firstworking

```
ffmpeg -i big_buck_bunny_720p.mp4 -ss 10 -t 10 bbb.y4m
target/release/rav1e samples/bbb.y4m --tile-rows-log2 1 --tile-cols-log2 1 -o bbb_tiled3.ivf
ffmpeg -i bbb_tiled3.ivf -c:v copy bbb_tiled3.mp4
```

:tada: