@rom1v 2019-04-05
> Would par_iter() keeps the run order of encode_tile()?

No, the point is to execute them in parallel.

> I.e. Preserving the order of generated bitstream from each tile.

Each encode_tile() returns a byte array (`Vec<u8>`), which are written in order (following the AV1 spec) at the end.

See 806005fb51f906ed92ed407ea55f213a91e615e2.