@rom1v 2019-04-05
> Would 'rec' covers a tile only, it comes from : "let rec = &mut ts.rec.planes[p];"
> I read that it still covers whole frame (plane p) ?

`ts` is a `TileStateMut<'_, T>`, `ts.rec` is a `TileMut<'_, T>`, `ts.rec.planes[p]` is a `PlaneRegion<'_, T>`. It only gives access to the plane region covered by the tile.