@rom1v 2019-04-06
Unfortunately (but as expected), the tiling structures are not a zero cost abstraction. They add an overhead in encoding time between 1~3%.

Concretely, if we compare the version compiled from `master` branch with the version compiled from this pull request, and encode a video without tiling (1 tile), the latter takes a bit more time to encode. Said otherwise, we pay a (small) cost for tile encoding even if we don't use it.

As an example, on my laptop, an encoding takes 3mn33,245 on `master` and 3mn35,722 on `tiling` (it's just an example, but the difference is quite stable).

You can compare encoding times on AWCY for 1 tile: https://beta.arewecompressedyet.com/?job=master-70005e353aa8ce21e3ecd257c927f71d4012a117&job=%402019-04-05T21%3A29%3A59.430Z_ref_1_tile

Maybe some work could be done to minimize this overhead (for example using more `#[inline(always)]`), but I think the overhead cannot be removed (within a memory-safe language).