@EwoutH 2019-04-08
I did a [scaling test](https://docs.google.com/spreadsheets/d/1JpPz-Tvj8OO9VGyv-0_VQQRlMVpChuwYeRwr3L3Ss1E/edit#gid=1605614229) with 1 to 32 tiles measuring speed, bitrate and psnr on a Ryzen 5 1600 (6c/12t, AVX2):

| Tiles |   FPS | Bitrate |   PSNR |
|-------|------:|--------:|-------:|
| 0x0   | 0,042 |    2443 | 50,757 |
| 0x1   | 0,078 |    2566 | 50,750 |
| 1x1   | 0,111 |    2586 | 50,734 |
| 1x2   | 0,103 |    2655 | 50,730 |
| 2x2   | 0,072 |    2686 | 50,727 |
| 2x3   | 0,046 |    2827 | 50,727 |

And normalized:

| Tiles |  FPS | Bitrate |    PSNR |
|-------|-----:|--------:|--------:|
| 0x0   | 100% |  100,0% | 100,00% |
| 0x1   | 186% |  105,0% |  99,99% |
| 1x1   | 264% |  105,8% |  99,95% |
| 1x2   | 245% |  108,7% |  99,95% |
| 2x2   | 171% |  109,9% |  99,94% |
| 2x3   | 110% |  115,7% |  99,94% |