@rom1v 2019-04-08
@EwoutH Thank you for your tests!

The number of threads is limited to 8 (on this branch, 4 on master). For 2×2 and 2×3 log2-tiles (4×4 tand 4×8 tiles), did you increase the number of threads? (option `--threads`)