@EwoutH 2019-04-08
@rom1v No I didn't, now rerunning with `--threads 12` for 1x1 to 2x3.

I'm now also testing on a 2x Intel Xeon E5-2680v4  (2x 14c/28t, AVX2) system, which should give better scaling results for 8 to 32 tiles. Results are coming in [here](https://docs.google.com/spreadsheets/d/1JpPz-Tvj8OO9VGyv-0_VQQRlMVpChuwYeRwr3L3Ss1E/edit#gid=1605614229).