@EwoutH 2019-04-08
Updated results.

2160p 8-bit 4:2:0 on a Ryzen 5 1600 (6c/12t, AVX2) system:

Tiles | FPS | Bitrate | PSNR | FPS | Bitrate | PSNR
-- | -- | -- | -- | -- | -- | --
0x0 | 0,042 | 2443 | 50,757 | 100% | 100,0% | 100,00%
0x1 | 0,078 | 2566 | 50,750 | 186% | 105,0% | 99,99%
1x1 | 0,113 | 2586 | 50,734 | 269% | 105,8% | 99,95%
1x2 | 0,114 | 2655 | 50,730 | 271% | 108,7% | 99,95%
2x2 | 0,077 | 2686 | 50,727 | 183% | 109,9% | 99,94%
2x3 | 0,050 | 2827 | 50,727 | 119% | 115,7% | 99,94%

1080p 8-bit 4:2:0 on a 2x Intel Xeon E5-2680v4  (2x 14c/28t, AVX2) system:

| Tiles |   FPS | Bitrate |    PSNR | FPS  | Bitrate | PSNR    |
|-------|------:|--------:|--------:|------|---------|---------|
| 0x0   | 0,103 |  722,02 |  49,680 | 100% | 100,0%  | 100,00% |
| 0x1   | 0,184 |  825,25 |  49,694 | 179% | 114,3%  | 100,03% |
| 1x1   | 0,311 |  835,29 | 49,6932 | 302% | 115,7%  | 100,03% |
| 1x2   | 0,465 |  875,76 | 49,6631 | 451% | 121,3%  | 99,97%  |
| 2x2   | 0,431 |  917,09 | 49,6699 | 418% | 127,0%  | 99,98%  |
| 2x3   | 0,418 | 1017,15 | 49,6727 | 406% | 140,9%  | 99,98%  |