@rom1v 2019-04-08
Thanks!

Did you use 16 threads (or more) for 2x2 (4x4) and 32 threads (or more) for 2x3 (4x8)?