@EwoutH 2019-04-08
@rom1v On Ryzen 5 12 threads for 1x1 to 2x3, on Intel Xeon 16 threads for 1x1, 32 for 1x2 and 56 for 2x2 and 2x3. CPU was maxed out on Ryzen.