@ycho 2019-04-08
With intra frame only, i.e. 1 frame only, the loss of coding efficiency is much less.

[AWCY for 1 frame, 1 tile vs 4 tiles](https://beta.arewecompressedyet.com/?job=%402019-04-05T21%3A29%3A59.430Z_ref_1_tile_1f%402019-04-08T18%3A16%3A54.523Z&job=%402019-04-05T21%3A28%3A52.445Z_4_tiles_1f%402019-04-08T18%3A17%3A34.602Z)

With largest with shield, >30%.