@rom1v 2019-04-08
@ycho Note that the rate distortion curves (at least when there are several frames) are meaningless until the corrupted frames bugs are solves (this [result](https://user-images.githubusercontent.com/543275/55742123-3aebb780-5a2f-11e9-95dd-29d0a2beefb5.jpg) has necessarily a poor psnr).