@tdaede 2019-04-08
That speed loss isn't so bad and worth it (the same infra will thread lrf etc too), but I think we could probably do a bit better. Do you have an idea where it's coming from? Could it be measured on a per-commit basis to narrow it down? Wouldn't surprise me if it could be eliminated by some strategic #[inline]s.