@ycho 2019-04-09
I think I've fixed "Issue 3: assertion failure at speed 0". 
Apply:
https://github.com/xiph/rav1e/pull/1196/commits/df716da15e43a8f60017f7704481ae82046fe4bc
(Or wait for PR #1196, which has this commit,  lands)

Turns out we've still have problems with directional intra prediction! And this was also the cause for my intar tx partitioning at depth 2.