@rom1v 2019-04-09
@ycho I confirm that https://github.com/xiph/rav1e/commit/df716da15e43a8f60017f7704481ae82046fe4bc prevents the assertion to fali.

However, can you make it fail on current `master` in some way? If not, chances are that tiling code break something, what do you think?