@EwoutH 2019-04-09
I ran an additional 3 runs and averaged the results out, reducing the noise a little.

| FPS | master | tiling | tiling.108 |
|-----|-------:|-------:|-----------:|
| 0x0 |  0,103 |  0,103 |      0,103 |
| 1x1 |      - |  0,304 |      0,308 |

| User time (s) | master | tiling | tiling.108 |
|---------------|-------:|-------:|-----------:|
| 0x0           |  705,4 |  719,0 |      719,2 |
| 1x1           |      - |  856,4 |      861,4 |

- There is no noticeable difference in single-tile performance (fps), but the used CPU time increase a little (about 2%), which would cause a small slow-down on a single-core system.
- It looks like the tiling.108 branch uses a little more cpu time (about 0.5%) with 4 tiles than the tiling branch, but the performance impact is inconclusive.
- [Full data here](https://docs.google.com/spreadsheets/d/1JpPz-Tvj8OO9VGyv-0_VQQRlMVpChuwYeRwr3L3Ss1E/edit#gid=350324363).