@rom1v 2019-04-18
Here are the new results after the recent changes on the `tiling` branch.

I tested on my laptop (8 CPUs) several encodings to compare encoding performance (this is not a good benchmark, but it gives an idea, you are encouraged to run your own tests), and launched AWCY runs.

## 1 tile

Since commits `Reduce overhead of tiling abstraction` (currently https://github.com/xiph/rav1e/pull/1126/commits/0d9e3dd7b39b93b46491b1a65d744950d5579d33) and `Inline all TileBlocks methods ` (https://github.com/xiph/rav1e/pull/1126/commits/43e2a566b576cac38707f3fd845bd1907286f6c7), with 1 tile (i.e. no tile encoding), `tiling` is similar / a bit faster than `master` now.

On this specific test:

```
time target/release/rav1e \
    Netflix_Crosswalk_1920x1080_60fps_8bit_420_60f.y4m \
    -o video.ivf
```

 - `master`: 7m5,630s
 - `tiling`: 7m2,336s (speedup: **1.008x**)

The difference happens to be visible on [AWCY](https://beta.arewecompressedyet.com/?job=tiling.137_reference_master%402019-04-18T18%3A42%3A18.213Z&job=tiling.137_1_tile%402019-04-18T17%3A42%3A49.905Z) (Metric: Encoding Time). This proves nothing since it probably depends on the server load, but the difference is quite consistant across different runs.

They produce the same results:

```
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.0000 |  0.0000 |  0.0000 |   0.0000 | 0.0000 |  0.0000 |     0.0000
```

## 2 tiles

```
time target/release/rav1e \
    Netflix_Crosswalk_1920x1080_60fps_8bit_420_60f.y4m \
    -o video.ivf
    --tile-cols-log2 1
```

 - 1 tile: 7m2,336s
 - 2 tiles: 3m53,578s (speedup: **1.81x**)

[AWCY](https://beta.arewecompressedyet.com/?job=tiling.137_1_tile%402019-04-18T17%3A42%3A49.905Z&job=tiling.137_2_tiles%402019-04-18T17%3A43%3A20.621Z) without low latency:

```
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
2.6419 |  2.4644 |  2.4862 |   2.6375 | 2.7437 |  2.7179 |     2.8091
```

[AWCY](https://beta.arewecompressedyet.com/?job=tiling.137_1_tile_low_latency%402019-04-18T18%3A41%3A43.143Z&job=tiling.137_2_tiles_low_latency%402019-04-18T18%3A31%3A33.779Z) with low latency (`--low_latency=true`):

```
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
1.3688 |  0.7373 |  1.1210 |   1.3709 | 1.4007 |  1.4120 |     1.3294
```


## 4 tiles

```
time target/release/rav1e \
    Netflix_Crosswalk_1920x1080_60fps_8bit_420_60f.y4m \
    -o video.ivf
    --tile-cols-log2 1 --tile-rows-log2 1
```

 - 1 tile: 7m2,336s
 - 4 tiles: 2m12,995s (speedup: **3.18x**)

[AWCY](https://beta.arewecompressedyet.com/?job=tiling.137_1_tile%402019-04-18T17%3A42%3A49.905Z&job=tiling.137_4_tiles%402019-04-18T17%3A43%3A55.951Z) without low latency:

```
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
5.9145 |  5.3288 |  5.8706 |   5.8448 | 6.1259 |  6.0573 |     6.0141
```

[AWCY](https://beta.arewecompressedyet.com/?job=tiling.137_1_tile_low_latency%402019-04-18T18%3A41%3A43.143Z&job=tiling.137_4_tiles_low_latency%402019-04-18T18%3A31%3A54.921Z) with low latency (`--low_latency=true`):

```
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
2.9473 |  1.8566 |  2.4672 |   2.9412 | 3.0758 |  3.0709 |     2.9937
```


## 8 tiles

```
time target/release/rav1e \
    Netflix_Crosswalk_1920x1080_60fps_8bit_420_60f.y4m \
    -o video.ivf
    --tile-cols-log2 2 --tile-rows-log2 1
```

 - 1 tile: 7m2,336s
 - 8 tiles: 1m57,533s (speedup: **3.59x**)

[AWCY](https://beta.arewecompressedyet.com/?job=tiling.137_1_tile%402019-04-18T17%3A42%3A49.905Z&job=tiling.137_8_tiles%402019-04-18T17%3A44%3A15.874Z) without low latency:

```
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
10.4391 |  9.9325 | 10.3241 |  10.2902 | 10.8110 | 10.7126 |    10.7646
```

[AWCY](https://beta.arewecompressedyet.com/?job=tiling.137_1_tile_low_latency%402019-04-18T18%3A41%3A43.143Z&job=tiling.137_8_tiles_low_latency%402019-04-18T18%3A32%3A21.077Z) with low latency (`--low_latency=true`):

```
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
5.1544 |  4.1440 |  4.6347 |   5.1364 | 5.3689 |  5.3737 |     5.2127
```